Adjusted takenAmount observation
-Added float cast when creating takenAmount observation to properly generate ratio
-Lowered distance penalization

+++ Assets/ThesisProject/Scripts/EnemyAgent.cs
@@ -198,7 +198,7 @@ public class EnemyAgent : Agent
                         distanceFromPlayer = distanceFromPlayer,
                         distanceFromBase = distanceFromBase,
 
-                        takenAmount = resourceObjectScript.totalDeposited / resourceObjectScript.dropAmount
+                        takenAmount = resourceObjectScript.totalDeposited / (float)resourceObjectScript.dropAmount
 
                     });
                     resourceCounter++;
